Fully working link between Pi and Pico.
The Python script (led_controller.py) on the Pi 4 formats a buffer of 60 LED commands,
which is received by the PIO-based RX-only SPI interface on the Pico, which then
encodes it into APA102 packets and sends them to the strip using that PIO state machine.
It is a wondrous thing to behold.