Conversion to C++
We're going to need a way to drive multiple strips, so it's time for the APA102 library
to become a class. SPI RX is commented out in this version, just to get things running.
Something is wonky with the size of the DMA transfer, and I'm sure it's just waiting
for its time to strike, but for now I'm just subtracting a 32-bit word from the
transfer and everything is fine. This bug waits for me...